Review updates

+++ cuda/base/batch_struct.hpp
@@ -66,7 +66,7 @@ inline gko::batch_multi_vector::uniform_batch<const cuda_type<ValueType>>
 get_batch_struct(const BatchMultiVector<ValueType>* const op)
 {
     return {as_cuda_type(op->get_const_values()), op->get_num_batch_entries(),
-            op->get_common_size()[1],
+            static_cast<int>(op->get_common_size()[1]),
             static_cast<int>(op->get_common_size()[0]),
             static_cast<int>(op->get_common_size()[1])};
 }
@@ -79,7 +79,7 @@ inline gko::batch_multi_vector::uniform_batch<cuda_type<ValueType>>
 get_batch_struct(BatchMultiVector<ValueType>* const op)
 {
     return {as_cuda_type(op->get_values()), op->get_num_batch_entries(),
-            op->get_common_size()[1],
+            static_cast<int>(op->get_common_size()[1]),
             static_cast<int>(op->get_common_size()[0]),
             static_cast<int>(op->get_common_size()[1])};
 }
